Setting the old slice position script as Legacy. Moving the new slice position algorithm as the ACR slice position task.
