+++ tests/test_acr_slice_position.py
@@ -11,8 +11,8 @@ class TestACRSlicePositionSiemens(unittest.TestCase):
     ACR_DATA = pathlib.Path(TEST_DATA_DIR / "acr" / "Siemens")
     slice_1_x_pts = [123, 129]
     slice_11_x_pts = [123, 129]
-    slice_1_y_pts = [40, 82]
-    slice_11_y_pts = [44, 82]
+    slice_1_y_pts = [45, 83]
+    slice_11_y_pts = [45, 83]
     dL = -0.59, -1.56
 
     def setUp(self):
@@ -22,19 +22,17 @@ class TestACRSlicePositionSiemens(unittest.TestCase):
         self.dcm_1 = self.acr_slice_position_task.ACR_obj.slice_stack[0]
         img_1 = self.dcm_1.pixel_array
         cxy_1, _ = self.acr_slice_position_task.ACR_obj.find_phantom_center(img_1, self.acr_slice_position_task.ACR_obj.dx, self.acr_slice_position_task.ACR_obj.dy)
-        mask_1 = self.acr_slice_position_task.ACR_obj.get_mask_image(img_1, cxy_1)
         self.slice1_x_pts, self.slice1_y_pts = self.acr_slice_position_task.find_wedges(
-            img_1, mask_1
+            img_1, cxy_1
         )
 
         self.dcm_11 = self.acr_slice_position_task.ACR_obj.slice_stack[-1]
         img_11 = self.dcm_11.pixel_array
         cxy_11, _ = self.acr_slice_position_task.ACR_obj.find_phantom_center(img_11, self.acr_slice_position_task.ACR_obj.dx, self.acr_slice_position_task.ACR_obj.dy)
-        mask_11 = self.acr_slice_position_task.ACR_obj.get_mask_image(img_11, cxy_11)
         (
             self.slice11_x_pts,
             self.slice11_y_pts,
-        ) = self.acr_slice_position_task.find_wedges(img_11, mask_11)
+        ) = self.acr_slice_position_task.find_wedges(img_11, cxy_11)
 
     # IMAGE 1
     def test_find_wedge_slice1_x(self):
@@ -68,8 +66,35 @@ class TestACRSlicePositionSiemens(unittest.TestCase):
 
 class TestACRSlicePositionGE(TestACRSlicePositionSiemens):
     ACR_DATA = pathlib.Path(TEST_DATA_DIR / "acr" / "GE")
-    slice_1_x_pts = [246, 257]
-    slice_11_x_pts = [246, 257]
-    slice_1_y_pts = [84, 164]
-    slice_11_y_pts = [89, 162]
-    dL = 0.41, 0.3
+    slice_1_x_pts = [247, 257]
+    slice_11_x_pts = [246, 256]
+    slice_1_y_pts = [94, 168]
+    slice_11_y_pts = [95, 169]
+    dL = 0.51, 0.41
+
+
+class TestACRSlicePositionPhilipsAchieva(TestACRSlicePositionSiemens):
+    ACR_DATA = pathlib.Path(TEST_DATA_DIR / "acr" / "PhilipsAchieva")
+    slice_1_x_pts = [126, 132]
+    slice_11_x_pts = [123, 129]
+    slice_1_y_pts = [44, 82]
+    slice_11_y_pts = [43, 81]
+    dL = 0.78, -3.12
+
+
+class TestACRSlicePositionSiemensSolaFit(TestACRSlicePositionSiemens):
+    ACR_DATA = pathlib.Path(TEST_DATA_DIR / "acr" / "SiemensSolaFit")
+    slice_1_x_pts = [124, 130]
+    slice_11_x_pts = [123, 129]
+    slice_1_y_pts = [44, 82]
+    slice_11_y_pts = [43, 81]
+    dL = -0.2, -1.37
+
+
+class TestACRSlicePositionSiemensLargeSliceLocationDelta(TestACRSlicePositionSiemens):
+    ACR_DATA = pathlib.Path(TEST_DATA_DIR / "acr" / "SiemensLargeSliceLocationDelta")
+    slice_1_x_pts = [123, 129]
+    slice_11_x_pts = [123, 129]
+    slice_1_y_pts = [41, 79]
+    slice_11_y_pts = [41, 79]
+    dL = -6.64, -6.84
